requirements.txt
streamlit
torch
torchvision
facenet-pytorch
Pillow
numpy
git+https://github.com/sb-ai-lab/EmotiEffLib.git
opencv-python-headless
